slet virker tilmeld virker ikke
